airflow/requirements.txt
certifi==2023.7.22
clickhouse-connect==0.6.14
lz4==4.3.2
minio==7.1.17
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
urllib3==2.0.6
zstandard==0.21.0
